Refactoring import declarations
I've decided that I really prefer module dependencies to be listed at the top of the file.  Imports of methods and values can occur inside lower scopes, but I generally prefer the class/object/package object dependencies to be at the top--not imported inside scopes or referenced with fully-qualified names at the use site.