Trying harder to grab submission metadata
Ugh.  Play is kind of an idiot when it comes to extracting parameters from jQuery requests made of JS objects--or maybe it's trying to be too smart?  Either way, it flattens out the parameters into an odd-to-query-for format, which was breaking submissions that came from jQuery.  I now try to be a bit smarter, myself, about trying to read the metadata.