Fixing obnoxious bug in `FileManager`
This bug has apparently been laying latent here for over a year.  How
nice!

Basically, it's something screwy with how Akka looks up actors.  You
give the actor a name with `system.actorOf(<blah>, name)`, but retrieve
it by name with `system.actorFor("/user/" + name)`, as if that makes any
sense....

Then, Akka's way of telling you that made the mistake of not
prefixing the name with `"/user/"` is to tell you that the actor you're
looking for "had already been terminated."  Thanks, Akka!