Changing determination of destinations of signed '.jar's
Previously, the output directory was being read from a configuration
property.  However, this ends up not working out so nicely.  In order to
be sufficiently general, the auto-signer ought to be able to output a
signed '.jar' to any location, as determined on the fly.  The property
wasn't all that useful, anyway....