Automatically signing JNLPs
NOTE: JNLPs that use a `codebase` attribute other than the default will
not be signed, for some fairly obvious security reasons.

This pre-emptively fixes an issue with Java, where it will soon be
disallowed to distribute unsigned JNLPs.  With the auto-signing in
place, we can continue to distribute our diversely-configured
WebStart applications as we have before, provided we use a signing
certificate from a root certificate authority.