Throwing out SBT logs every few minutes
To accomplish this, we find the SBT log file associated with the current
run of SimServer and repeated clear it.

We can't delete it, because the Play process keeps a handle on it, and
reducing log levels doesn't help the log file from gathering up an
obscene amount of data--so much data, in fact, that it threatens the
stability of the ABMPLUS server by filling up the disk drive.

We could probably get around this problem by running the server in
production mode, but that would break all of the dynamic routing that
we do.  So, instead, let's just repeatedly wipe it clean.