Introducing the unsafe HubNet registry!
This sucks.  Not that the old system was really very secure, but new
constraints within the project have essentially forced me to throw out
anything with a semblance of security.  This new system allows for
teachers to register their IP/ports without using encryption.  The was
also some refactoring to get the old system and the new one sharing some
code.  Definitely not my most wonderful work....